refactor: 기존 재고 차감 dto 위치 변경 (#128)
- facade -> service
