+++ src/main/java/com/example/i_commerce/domain/product/application/dto/StockDeductCommand.java
@@ -1,5 +1,8 @@
-package com.example.i_commerce.domain.product.facade.dto;
+package com.example.i_commerce.domain.product.application.dto;
 
+import lombok.Builder;
+
+@Builder
 public record StockDeductCommand(
     Long productItemId,
     int quantity,
