+++ src/test/java/com/example/i_commerce/domain/product/service/StockServiceTest.java
@@ -14,7 +14,7 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 
 import com.example.i_commerce.domain.product.application.service.StockService;
-import com.example.i_commerce.domain.product.facade.dto.StockDeductCommand;
+import com.example.i_commerce.domain.product.application.dto.StockDeductCommand;
 import com.example.i_commerce.domain.product.entity.Stock;
 import com.example.i_commerce.domain.product.entity.StockChangeType;
 import com.example.i_commerce.domain.product.entity.StockHistory;
